+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2025-07-15
+
+### Added
+- CPU temperature monitoring
+- Toggle temperature display option
+- Improved menu bar display
+
 ## [1.0.0] - 2025-07-15
 
 ### Added